Maps in results done - takes lat/long from DB displaying corectly on mob
