+++ database.sql
@@ -308,3 +308,22 @@ insert into booking (id, property_id, guest_id, date_booked, date_start, date_en
 insert into booking (id, property_id, guest_id, date_booked, date_start, date_end) values
   (2, 6, 1, '2018-10-21T10:37:33.735972+01:00', '12-Dec-2018', '16-Dec-2018');
 alter sequence booking_id_seq restart with 3 increment by 1;
+
+
+-- ALTER TABLE property
+  -- ADD COLUMN lat decimal
+
+-- ALTER TABLE property
+  -- ADD COLUMN lng decimal
+
+-- UPDATE property
+--   SET lat = -33.8710844, lng = 151.2233585
+--     WHERE id = 8
+
+-- UPDATE property
+--   SET lat = 33.8687895, lng = 151.1942171
+--     WHERE id = 9
+
+-- UPDATE property
+--   SET lat = -33.8708464, lng = 151.20733
+--     WHERE id = 10
