--- src/components/map.js
@@ -1,64 +1,0 @@
-import React, { Component } from 'react'
-import {Map, InfoWindow, Marker, GoogleApiWrapper} from 'google-maps-react';
-const APIKey = 'AIzaSyBHm1z95Or8WefNxOjZ-wejrcZqEcRkVwY'
-const AnyReactComponent = ({ text }) => <div>{ text }</div>;
-
-class MapView extends Component {
-
-  // static defaultProps = {
-  //   center: { lat: 40.7446790, lng: -73.9485420 },
-  //   zoom: 11
-  //   }
-    constructor(){
-      super()
-    }
-
-  render() {
-      return (
-        <Map google={this.props.google} zoom={9}
-            style={{ height: '250px', width: '100%' }}
-            initialCenter={{
-            lat: 40.854885,
-            lng: -88.081807
-          }}>
-
-          <Marker onClick={this.onMarkerClick}
-                  title={'Current location'} />
-
-          <InfoWindow onClose={this.onInfoWindowClose}>
-              <div>
-                <h1>London</h1>
-              </div>
-          </InfoWindow>
-
-        </Map>
-      );
-    }
-  }
-
-export default GoogleApiWrapper({
-  apiKey: (APIKey)
-})(MapView)
-
-
-
-// take an address from State
-// fetch to the google APi with it
-// return the long lat
-//that can then go into maps - stage 2
-
-
-
-function googleGeocoding(postCode, APIKey){
-  fetch(`https://maps.googleapis.com/maps/api/geocode/json?address=${postCode}&key=${APIKey}`)
-  .then(function(result){
-    return result.json()
-  })
-  .then(function(data){
-    console.log(data)
-    console.log(data.results[0].location)
-    console.log(data.results[0].location.lat)
-    console.log(data.results[0].location.lng)
-    //set state with the lat long maybe? Or call this function inside our "proper" GMaps function
-  })
-}
